+++ src/gcp/frameworks.ts
@@ -1,12 +1,77 @@
 import { Client } from "../apiv2";
 import { frameworksOrigin } from "../api";
-import {
-  BuildOutputOnlyFields,
-  Operation,
-  Stack,
-  Build,
-  StackOutputOnlyFields,
-} from "../frameworks/compose/api/interfaces";
+
+type State = "BUILDING" | "BUILD" | "DEPLOYING" | "READY" | "FAILED";
+
+interface Codebase {
+  repository?: string;
+  rootDirectory: string;
+}
+
+/** A Stack, the primary resource of Frameworks. */
+export interface Stack {
+  name: string;
+  mode?: string;
+  codebase: Codebase;
+  labels: Record<string, string>;
+  createTime: string;
+  updateTime: string;
+  uri: string;
+}
+
+export type StackOutputOnlyFields = "createTime" | "updateTime" | "uri";
+
+export interface Build {
+  name: string;
+  state: State;
+  error: Status;
+  image: string;
+  source: BuildSource;
+  buildLogsUri: string;
+  createTime: Date;
+  updateTime: Date;
+  sourceRef: string;
+}
+
+export type BuildOutputOnlyFields = "createTime" | "updateTime" | "sourceRef";
+
+interface BuildSource {
+  codeBaseSource?: CodebaseSource;
+}
+
+interface Status {
+  code: number;
+  message: string;
+  details: any[];
+}
+
+interface CodebaseSource {
+  // oneof reference
+  branch: string;
+  commit: string;
+  tag: string;
+  // end oneof reference
+}
+
+interface OperationMetadata {
+  createTime: string;
+  endTime: string;
+  target: string;
+  verb: string;
+  statusDetail: string;
+  cancelRequested: boolean;
+  apiVersion: string;
+}
+
+export interface Operation {
+  name: string;
+  metadata?: OperationMetadata;
+  done: boolean;
+  // oneof result
+  error?: Status;
+  response?: any;
+  // end oneof result
+}
 
 export const API_VERSION = "v2";
